Combine both almost_completed files into the same file
