--- backend/hoagieplan/api/dashboard/almost_completed_api.py
@@ -1,45 +1,0 @@
-from typing import List, Dict
-
-from django.http import JsonResponse
-from rest_framework.decorators import api_view
-
-from hoagieplan.api.dashboard.almost_completed_reqs import (
-    get_all_program_data,
-)
-from constants import MINORS, CERTIFICATES
-from hoagieplan.logger import logger
-
-
-@api_view(["GET"])
-def almost_completed(request):
-    """Return a JSON list of almost-completed programs for the current user.
-
-    Response format: [{"code": "COS", "name": "Computer Science", "needed": 1, "type": "minor", "prereqFulfilled": true|false|null, "independentWorkRequired": true|false, "incompleteRequirements": ["Electives (1 more)", "Advanced Courses (2 more)"]}, ...]
-    """
-    try:
-        # Optimized: compute all data in a single check_user() call
-        results, prereq_status, iw_status, incomplete_subreqs = get_all_program_data(request.user)
-
-        # results is dict{code: needed_count}. Convert to list with names from MINORS/CERTIFICATES
-        out: List[Dict] = []
-        for code, needed in results.items():
-            name = MINORS.get(code) or CERTIFICATES.get(code) or code
-            typ = "minor" if code in MINORS else ("certificate" if code in CERTIFICATES else "unknown")
-            prereq_fulfilled = prereq_status.get(code)  # Can be True, False, or None
-            independent_work_required = iw_status.get(code, False)  # Default to False if not found
-            incomplete_requirements = incomplete_subreqs.get(code, [])
-            
-            out.append({
-                "code": code,
-                "name": name,
-                "needed": needed,
-                "type": typ,
-                "prereqFulfilled": prereq_fulfilled,
-                "independentWorkRequired": independent_work_required,
-                "incompleteRequirements": incomplete_requirements
-            })
-
-        return JsonResponse({"programs": out})
-    except Exception as e:
-        logger.error(f"Failed to compute almost completed programs: {e}", exc_info=True)
-        return JsonResponse({"programs": []})
